moves v3 services to root level
