+++ app/Http/Controllers/ReviewsController.php
@@ -4,7 +4,7 @@ namespace Rogue\Http\Controllers;
 
 use Rogue\Models\Post;
 use Illuminate\Http\Request;
-use Rogue\Services\Three\PostService;
+use Rogue\Services\PostService;
 use Rogue\Http\Controllers\Legacy\Two\ApiController;
 use Rogue\Http\Transformers\PostTransformer;
