+++ app/Jobs/ImportTurboVotePosts.php
@@ -7,9 +7,9 @@ use League\Csv\Reader;
 use Rogue\Models\Post;
 use Rogue\Models\Signup;
 use Illuminate\Bus\Queueable;
-use Rogue\Services\Three\PostService;
+use Rogue\Services\PostService;
 use Illuminate\Support\Facades\Storage;
-use Rogue\Services\Three\SignupService;
+use Rogue\Services\SignupService;
 use Illuminate\Queue\InteractsWithQueue;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Foundation\Bus\Dispatchable;
